Create multiplicativePersistence
6kyu #6
Takes in an integer and returns the number of times the digits of that number need to be sequentially multiplied in order to reach 1 digit.
I struggled with this one a lot. The code itself took me not that long, but then figuring out that it was adding 48 to my variable for no reason took several hours. I have yet to figure out why, but in the mean time, I'll just counteract that process by subtracting 48 in line 13.